Rename most Intersection* samples to CaptureConverted*.
Although the JLS specifies that capture conversion produces intersection
types, some tools may handle such situations differently internally.

(For the remaining Intersection* class, IntersectionSupertype, I'm not
entirely sure whether it would technically deal in intersection types or
not. But it's at least a bit different than these other tests, so I am
happy to leave it with a distinctive name. If we find a better name
later, we can rename it, too.)