Merge fullyQualified classes into fewer packages.
This may help with the problems in discussed in
https://github.com/jspecify/nullness-checker-for-checker-framework/pull/4

- https://github.com/jspecify/nullness-checker-for-checker-framework/pull/4#pullrequestreview-562049162
- https://github.com/jspecify/nullness-checker-for-checker-framework/pull/4#issuecomment-755657417

However, my guess is that it's not completely sufficient, since we'll
still have sources in 2 different directories (albeit one nested inside
the other). I'll test and see what happens.

As part of this change, I noticed that the directory structure did not
quite match the Java packages of its classes: The directory structure
had a fullyqualified directory underneath the parent fullyQualified
directory. That child directory should have appeared in the package
name, but it did not. I solved this by adding it to the package name at
the same time that I removed other components from the package name.