Sample for casting from a wildcard capture to a type variable.
The NullableSupplier part of this sample is working in our prototype
(after
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/559ba69c20d5e2f2b828032ee82a996f3a949eee)
but really only by accident: I don't think there ought to be anything
wrong with casting the result to a `T` per se: The result of such a cast
ought to be a `@Nullable T`. Subsequently _returning_ a `@Nullable T`
_ought_ to be a problem. But our prototype is reporting the error at the
site of the _cast_, even if I separate the cast from the return. (I'm
pretty sure that it reports the error even if the method would be
satisfied with a `@Nullable T`, though I'd want to confirm.)

There is a larger question here of what a cast to `T` means: I wouldn't
expect it to require the value to be non-null (just as a cast to, e.g,.
`String` doesn't), and I'm unsure whether the Checker Framework views
this differently or we're just breaking the usual CF handling with our
non-standard treatment of type parameters and wildcards. (Certainly that
non-standard treatment is getting us into trouble in other places,
perhaps including in `isTypeCastSafe` itself, as noted in a new
implementation comment in the prototype.)