Add "trusted" the beginning of the "null-inclusive" and "null-exclusive" terms.
Yes, because we really need those terms to be longer... :)

It occurs to me that it might make sense to align our "least convenient
world" and "most convenient world" terminology with this idea of
"trust." Perhaps they'd becomes "doubting" and "trusting" or something?
But I also worry about confusion: It's not as if the "doubting" world
doesn't still "trust" sometimes.

(I also considered "assumed.")