Check `instanceof Foo` instead of `instanceof Object`.
The latter is entirely equivalent to a null check. Thus, if the result
is `false`, tools may want to recognize that the value is `null`.

Perhaps there should be a sample that checks just that, but `instanceof
Object` is rare enough that recognizing it is more of a "nice to have"
than an essential. (We already try to avoid assuming most dataflow logic
in our samples, but we've made exceptions for common cases that we
expect ~every tool to want to recognize.)

Fixes https://github.com/jspecify/jspecify/issues/164