If a value is equal to null, make it nullable, even if it was unspecified before.
We are "anti-refining" it. This doesn't normally make sense, but
unspecified nullness is weird.

Note that I *haven't* updated the samples that check that a *non-null*
value is equal to null. I probably should: Our knowledge that a value is
non-null often comes from a method's declared contract, which could be
violated but callers/implementations who don't run a sound nullness
checker or who suppress errors. In contrast, our knowledge that the
value *is* null is backed up by a runtime check. So it's possible for
the former to be wrong but not the latter.

(The above explanation is an expanded but possibly still unclear rewrite
of what I said in
https://github.com/jspecify/jspecify/issues/164#issuecomment-774120716)