Put sources jar, prefer Java 9+ sources over Java 8.
This is like https://github.com/jspecify/jspecify/pull/180 but for the
sources jar instead of the javadoc jar.

I didn't find built-in configuration similar to `javadoc { ... }` for
sources, but then again, I have no idea where to look :)

There must be a more idiomatic way to do this. And what I'm doing now
may even be outright wrong in some way. But I've spent close to 2 hours
on what I hoped would be a 10-minute project, so I'm going to lock in
the gains(?) I have so far and then give up.

Note that this commit also removes module-info.class from the sources
jar. I'm not sure why it's there now (or even if that's a normal thing
to do), but it seems weird. We can add it back easily if desired.

Closes https://github.com/jspecify/jspecify/issues/179