Redo mrjar handling.
This commit modifies the main source set to exclude DefaultNonNull and
then compensates by adding the main DefaultNonNull to the configuration
for the jar.

This is the possibility that I'd mentioned in
https://github.com/jspecify/jspecify/commit/57f79c69f4a19e045bc85548beb4074b5e3d272b#commitcomment-48004505

This may produce its own complications if we someday have dependencies
between a class with per-Java-version variants and a class without. If
we're lucky, the dependeny will run one way only, and we can just
configure the build to make the one compilation task depend on the
other.

(So much for giving up on this as I'd planned! At least I'm learning a
bit about Gradle.)