Eliminate separate compileJava8Java task.
I had nearly committed my earlier change in a form that used the wrong
--release for that task, so I'd rather avoid creating a task unless we
really need to.

This is probably a good policy in general: I wouldn't be surprised if
other built-in or extra Gradle tasks had special handling for
compileJava. Maybe even some tests try to run against the output of that
task, rather than against the jar? If so, we want to make sure that
compileJava covers everything. Then again, it doesn't cover
"everything," thanks to the second version of DefaultNonNull.... So we
may run into trouble in that case, anyway. Still, this commit seems like
an improvement, in part because it's a general simplification.