Release more JDK-related samples.
Contrary to my claim in the previous commit, they work fine: I was
seeing a problem only because I was accidentally mixing a repo in which
our annotation was named `@DefaultNonNull` with a repo in which it was
not.

(I still haven't included the Guava samples. To do that, I'll not only
need to ensure that Guava is on the classpath but also ensure that we
use a version of Guava that has been annotated for JSpecify. Possibly we
could pull one from a special branch in GitHub and built it ourselves,
but that doesn't seem worth the effort.)

I do note that the GitHub Actions CI is still failing. This is because
our CF fork has not been updated to contain the CF commit that
corresponds to
https://github.com/typetools/stubparser/commit/25e36946755b040d23c0af8f4b7bd2bc0fcf4042
I'm working on that update now.