Update file names for naming of NullMarked.
We'd been using the name "NullAware" here, since this work dates from
back when that was our best guess at the name.

I'd fixed most of the references in source files, but this commit covers
2 things I'd missed:

- I hadn't renamed the files to match.

- I had run a search-and-replace operation without `/g`, so I had
updated only the first occurrence of "NullAware"/"DefaultNonNull" in a
give line. I had thought this was fine, since those names typically
appear only once per line (as an import or a non-repeated annotation).
However, I'd neglected
NullAwareDirectUseOfNotNullAwareBoundedTypeVariable, which contains
"NullAware" twice. So I had to update that source file (in addition to
the rename I would have had to do regardless).