Say "null-marked scope" instead of "null-marked context."
I think this is my mistake from the 0.1 draft: In moving away from
"defaulting" terminology (since it appeared likely that we'd drop the
`@DefaultNonNull` name), I replaced it with "context," a term that
appeared with this meaning only once in the Draft Design Proposal doc.

But in fact we've been using "scope" for this most of the time, e.g.,
https://github.com/jspecify/jspecify/issues/154

Plus, we give the term "type context" has a particular meaning that is
different from the one here (e.g., in "non-nullable type context").