Remove the distinction between "illegal" locations and "unspecified" ones.
There is controversy about whether some of these locations should be
legal or not, and it's not clear that we gain anything by taking a
position.

Plus, "illegal" may sound a bit like a description of tool behavior
(though I'm not sure how people see it in general).

Plus, "unspecified" already means something to us, so it's nice to avoid
that.