Refer to "all worlds" and "some world" instead of "the least convenient world" and "the most convenient world."
This may help emphasize that a given type usage (e.g., a field type) can
be be treated as if it is `UNION_NULL` in one case and `NO_CHANGE` in
another, and each of those may be "convenient" in different cases. But
this commit isn't actually required for that, since the idea is already
that that field may have a different nullness operator each time we
refer to it.

Still, I had some reason for believing that the changes of this commit
would be a good thing a while back. Maybe part of my logic was that
there could be multiple worlds that were equally convenient (or
inconvenient)? Plus, "all"/"some" feels like the right model to
emphasize, even more so than the "adversarial" nature of "most/least
convenient." So let's go with that.