Remove "trusted" from the "null-* under every parameterization" terms.
This undoes bb3c6a6bf620724f58c363c4b93534395935fcca.

I'm not really sure why I liked the idea of "trusted" so much. I don't
really like it now: For any given "world," the type either is or isn't
null-* under every parameterization. The "trusted" term would make more
sense if we weren't already talking about multiple worlds -- that is, if
we were "trusting" that we're in the "right" world. It's not like we
"trust more" in one world than in another.

And this makes the term shorter.