Fix subtyping so that `T` is not necessarily a subtype of `T!!`.
It's, well, not :) For example, the following should be recognized as
unsound:

```java
class MyList<E extends @Nullable Object> {
  List<E> elements = ...;

  void add(E e) {
    elements.add(e);
  }

  @Nonnull E get(int i) {
    return elements.get(i); // unsound
  }
}
```

I should give credit to the Kotlin people, who already had this rule. I
had omitted it because I didn't understand it :(
https://groups.google.com/g/jspecify-dev/c/yPnkx_GSb0Q/m/evTRjAVvAQAJ

I'm not sure if the rule can actually come up in Kotlin (though I
shouldn't state anything with too much confidence now :)). In fact,
depending on exactly what we do with "defaulting" of type-variable
usages, I'm not sure it could come up in JSpecify-using Java code,
either:

The sample input in this commit demonstrates a way that it can arise
with the current spec, but that could change. And the example code
snippet above requires `@Nonnull`, which we don't currently offer (but
that, too, could change).

Anyway, this commit to the spec solves this problem for the case that
exists with the spec as it is today. If we introduce support for
`@Nonnull` in the future, we should think more about whether this change
is sufficient. It may well be; I just had the initial reaction that we'd
need to look for `MINUS_NULL` at every step of the nullness-establishing
path. I considered doing that even now, but it's a little tricky because
the path can be zero steps long. Thus, doing it "at every step" isn't
enough. And the more I thought about it, the less convinced I was that
it was actually required. Maybe it can come up in _some_ case, though,
like with `? super`? Or maybe it can even come up today....

Perhaps the lesson I should be learning here is to base my change more
directly on Kotlin's rules. The way to do that would be to write the
following instead of the new `MINUS_NULL` check:

> ...and `F` is _not_ null-exclusive under every parameterization.

But given that I can't prove the need for that (and that it may be
expensive to compute), I haven't done it in this commit.