Demonstrate some fallout from the `T <: T!!` fix.
(Also, fix a type in the class name / filename of the sample that I
included with that fix.)

```java
TypeVariableMinusNullVsTypeVariable.java:53: error: [nullness] incompatible types in return.
      return cast.caching();
                         ^
  type of expression: NonnullSupplier<T!!>!!
  method return type: Supplier<T>!!
```

(This happens in both strict mode and lenient mode.)

This code is based on
https://github.com/google/guava/blob/08e8d34e304efc3040ba528337d396aaa11638be/guava/src/com/google/common/collect/Lists.java#L803

As discussed in the code comment in the new sample, I think I'm actually
OK with making this an error. It's not a useful error here, but it's
justifiable, and I expect it to be rare. If it turns out to be more
common (and still not useful), we should reconsider: The bug fixed by
the previous commit was also unlikely to be tickled by real code,
affected only code that already produced other warnings, and prevented
only warnings, not errors.

Now, Guava does see *other* new failures from the previous change.
However, they aren't enough to get me to revert it:

https://github.com/google/guava/blob/08e8d34e304efc3040ba528337d396aaa11638be/guava/src/com/google/common/collect/Iterables.java#L380
becomes an error, but that seems to be another bug, which, as usual, I
would start by blaming on
https://github.com/jspecify/checker-framework/issues/4

https://github.com/google/guava/blob/08e8d34e304efc3040ba528337d396aaa11638be/guava/src/com/google/common/collect/Platform.java#L101
becomes _doubly_ an error, but again, there was already an error, and I
think that both are legitimately pointing out that I let my let the
annotations on `Arrays.copyOfRange` get out of sync with the annotations
there.

I want to at least test the previous commit more widely before backing
it out. If the problems run even deeper than I've realized, I'd like to
know.