Partially fix anchors/fragments/targets/slugs.
The `<a name="...">` approach has the problem that that tag is included
not just in the header itself but also in the table-of-contents entry
for that header. This means that links resolve to the ToC, so they go
essentially nowhere.

(_Plus_, the parser includes the HTML tags when it generates the anchor.
This results in anchors like
"#bound-of-an-unbounded-wildcard-a-name-unbounded-wildcard-a.")

The syntax I use in this commit is from
https://myst-parser.readthedocs.io/en/latest/syntax/syntax.html#targets-and-cross-referencing

This change still doesn't make clicking the "permalink" icon use _our_
preferred anchor name. Instead, it sticks to the autogenerated one. To
fix that, we'd probably need to follow these instructions:
https://myst-parser.readthedocs.io/en/latest/syntax/optional.html#auto-generated-header-anchors

That would let us hardcode the anchor text for any sections that we want
to.