Add a `developer` to the project metadata.
Instructions:
https://docs.gradle.org/current/userguide/publishing_maven.html#sec:modifying_the_generated_pom

Without this, we can't close the Sonatype repository: We get the error
"Developer information missing."

It's not clear who exactly we should list here. Probably we'd prefer
that people contact us on GitHub or our mailing list (if we've opened up
posting permission to everyone). For the moment, I've added only kevinb,
since he has plans to do some outreach to the wider community. If anyone
would like to be added, feel free to add yourself.