Remove design overview doc.
It's still available to the project group at
https://docs.google.com/document/d/15NND5nBxMkZ-Us6wz3Pfbt4ODIaWaJ6JDs4w6h9kUaY/edit

At last, though, we have the beginnings of
https://jspecify.dev/user-guide available to people who are looking for
an overview of the project. With that, the design overview would fit
best as an intermediate doc between that and the spec. However, we
haven't kept it up to date, so it currently clashes with the other two
docs. Ideally we'll fix that one day, or maybe we'll find intermediate
content to fit better in a FAQ, in individual design documents, etc.