Set version to 0.0.0-SNAPSHOT.
We released 0.2.0, so it's well past time for us to move past
0.1.0-SNAPSHOT.

The next obvious possibility would be 0.3.0-SNAPSHOT (or, if we're
optimistic, 1.0.0-SNAPSHOT). However, using _any_ version number
complicates matters for anyone who wants to build this repo and use it.
Notably, our own
https://github.com/jspecify/nullness-checker-for-checker-framework
suggests using jspecify-0.1.0-SNAPSHOT.jar on the classpath. (And that
same project has had mild difficulties with the Checker Framework test
jar, which has a name like framework-test-3.11.1-SNAPSHOT.jar:
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/edd8a769ef8c204d121ed8e39ba1a56e8b33aaca#diff-49a96e7eea8a94af862798a45174e6ac43eb4f8b4bd40759b5da63ba31ec3ef7)

To avoid this in some Google projects, we've used the version "number"
of "HEAD-SNAPSHOT":
https://github.com/google/truth/blob/3c580fc7971d85ba3b0b642310b29f101a7451c7/pom.xml#L14

However, that causes problems with the Java module system: Even though
[`ModuleDescriptor.Version`](https://docs.oracle.com/en/java/javase/16/docs/api/java.base/java/lang/module/ModuleDescriptor.Version.html)
suggests that "HEAD-SNAPSHOT" should be valid, the implementation
rejects it.

Other people have this problem:

- https://issues.apache.org/jira/browse/MCOMPILER-446
- https://github.com/AlmasB/FXGL/issues/1091#issuecomment-888832276
- https://github.com/devonfw/ide/issues/539#issuecomment-802142786

I liked the last link's suggestion of a "version 0," so here we are :)