Remove false claim about `ImmutableList.Builder<@Nullable Foo>`.
As I continually forget, `add(null)` would be an error not just for the
current special case (`MINUS_NULL`) but also for the hypothetical
alternative (`NO_CHANGE`).

There might be _some_ other statement about `ImmutableList.Builder`
that's true and at least a _little_ interesting. For example:

```
@NullMarked
<E extends @Nullable Object> go(ImmutableList.Builder<E> builder, E e) {
  builder.add(e);
}
```

I _think_ the `NO_CHANGE` approach would permit the `add` call. (But
it's worth noting that it would still reject the code because
`ImmutableList.Builder<E>` has an out-of-bounds type argument.) The
`MINUS_NULL` approach, by contrast, rejects it.

But it's not clear that that matters much in practice -- or that our
approach is better in the few cases in which it does matter! So, I've
substituted a claim that I think is both true _and_ more interesting
than this new `ImmutableList.Builder` example.

I've also removed the claim that we follow Kotlin precedent. Some quick
testing suggests that we do not. But I've gotten a little confused. So
I'm removing it not because I'm confident that it's false, just because
I'm no longer confident that it's true.