Loosen a condition of the lower-bound rule.
-   Before: "Its nullness operator is `NO_CHANGE`."
-   After: "Its nullness operator is *not* `MINUS_NULL`."

I think the need for this change is demonstrated by [the call to
`lib.useT(t)` in
`SuperTypeVariableUnspec.nullableBounded`](https://github.com/jspecify/jspecify/blob/80410e5e6a8780b9bc71a3c1133f65b061182d17/samples/SuperTypeVariableUnspec.java#L124):
Even in "strict mode," that call should be permitted. But given that...

-   `t` is not of a type that is null-exclusive under every
    parameterization and
-   `useT`'s parameter is not of a type that is null-inclusive under
    every parameterization (at least not in strict mode)

...we need the "path" rule to apply here. And if that rule requires the
captured type variable's bound to have an operator of `NO_CHANGE`, then
the rule doesn't apply in strict mode.

(At one point, I think I'd been assuming that we didn't need to address
the `UNION_NULL` and `UNSPECIFIED` cases because they would be covered
by the check for null-inclusive under every parameterization. That is
likely true of `UNION_NULL`. But it's not true of `UNSPECIFIED` _in
strict mode_. (Given that the problem looks to be restricted to
`UNSPECIFIED`, I could likely have instead changed this rule to
"`NO_CHANGE` or `UNSPECIFIED`. But that would have looked weird: It's
not that `UNION_NULL` would be _wrong_ to include, just not absolutely
_necessary_.))

(There was also a part of me that thought I could remove this bullet
entirely: I thought I could rely on the general nullness subtyping
condition of "`F` does not have nullness operator `MINUS_NULL`."
However, that's talking about a case in which a type like `interface
PredicateButActuallyRequiresNonNull<T extends @Nullable Object>` has a
method `boolean test(@NonNull T t)`. What we're talking about here is a
case in which we have a `Predicate<? super @NonNull T>` and we're trying
to pass a `T` to it. I think. So I think we still need the bullet, just
in this commit's modified form.)