Remove weak justification for why we picked `MINUS_NULL`.
The more I look at this, the less convincing I find it: Type inference
aside, it's already easy to see that the argument to `Optional.of` is
null-exclusive under every parameterization.

(We could well switch to `NO_CHANGE` if people prefer. Maybe I'll test
more to see what Kotlin does...)