Expand samples for ternary operator.
I'm debugging a false positive in code like the following:
https://github.com/google/guava/blob/156694066b5198740a820c6eef723fb86c054343/guava/src/com/google/common/base/Throwables.java#L470

As I'm working to merge the CF implementation of capture conversion into
our prototype checker, I'm seeing the checker infer the type of the
ternary as `Method` rather than the expected `Method?`. The result is an
error.

I've been looking at this for a little while, and I'm losing track of
what I've managed to reproduce under what conditions :)

It would have seemed that part of the story would have to be that we've
disabled `conditional` checks for our checker's tests by suppressing
them here:
https://github.com/jspecify/nullness-checker-for-checker-framework/blob/01e2a2b51d212579ff1bc3167f7e86a8399c6b09/src/test/java/tests/NullSpecTest.java#L38

(Note also that we have no handling for them here:
https://github.com/jspecify/nullness-checker-for-checker-framework/blob/01e2a2b51d212579ff1bc3167f7e86a8399c6b09/src/test/java/tests/NullSpecTest.java#L127)

In contrast, we leave `conditional` checks on when using our prototype
against "real code" internally.

However, even with `conditional` off, we still see all expected errors
here.

That makes some sense, given that such checks are theoretically
redundant:
https://github.com/jspecify/checker-framework/blob/750c071d2c987770964f6e687ffecda20c8e6f1d/framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java#L2164-L2167

However, after capture conversion, the story may be different: We might
need `conditional` on in order to detect real problems. Maybe that in
turn will introduce the spurious strict-mode errors that we'd sought to
avoid:
https://github.com/jspecify/jspecify/commit/d7b4fa5d5bb9b86db54c8334a960b6a717e75777#diff-73a7495a497652a6c0da68829b5f848455ebdf174798b0787c5cdbaf3a526ff4