Remove NullnessUnspecified.
We may add this again someday, but for now, we want to make sure that we
don't accidentally include it in a release :)

The annotation is still used in our sample inputs. To partially address
that, I'll restore the class in the branch samples-google-prototype,
which is the only branch that I believe anyone is actively testing
against.

If this causes problems for others, please let me know.