Remove TODO about substitution into type-variable bounds.
It's still conceivable that we need such a thing. But if we do, it's
likely more correctly addressed at a deeper level. Notably, I see that
[JLS
4.10.2](https://docs.oracle.com/javase/specs/jls/se11/html/jls-4.html#jls-4.10.2)
has no mention of substitution in its discussion of type-parameter
bounds.

Additionally, with our hacky test checker, I am finding that *undoing*
the Checker Framework's subsitution into type-variable bounds *improves*
results. *That said*, I remember some other change that improved results
(but maybe only on *net*?) not long ago that I decided to back out
because it seemed to be getting the right answers for the wrong reasons.
So I'm not sure that undoing substitution is *actually* right, though
it's possible.

Honestly, I can't quite get my head around "substitution into
type-variable bounds" at all today. Arguably I should leave the TODO for
a day when I'm feeling more up to it. But I'm going to remove it in the
hope that:

- The bounds at the declaration site always provide enough information,
  since they must always be as strict as any bounds at a use site.
- If there is "extra" information available from viewing the type
  variable "as a member of" a type, then we get that from the type
  argument supplied to that type (a type argument that may itself by a
  type variable) or perhaps wildcard capture?