Samples for arrays.
We are at least tentatively leaning toward encouraging tools to [embrace
array
covariance](https://github.com/jspecify/jspecify/issues/65#issuecomment-1027648171).
That's hardly settled, but we probably benefit from having samples for
it so that checker authors can use them to monitor their checkers'
behavior.

The original motivation for this commit was actually not the subtyping
rules but the _same-type_ rules: Those are the rules that our hacky
checker [_currently_](https://github.com/jspecify/jspecify/issues/49)
uses for parameters in overrides. And it turns out that our hacky
checker is performing little checking of array-element types, if any.
Specifically, it did not report the mismatch between
`ImmutableMultiset.copyIntoArray` and the method it overrides:

- https://github.com/google/guava/blob/a5eb7fd6fbc3e5fdebe65f90dc6c41729f17b7b0/guava/src/com/google/common/collect/ImmutableCollection.java#L383
- https://github.com/google/guava/blob/a5eb7fd6fbc3e5fdebe65f90dc6c41729f17b7b0/guava/src/com/google/common/collect/ImmutableMultiset.java#L328