Add JDK19 (non-Early Access) as a non-experimental CI job.
We expect that this works after
https://github.com/jspecify/jspecify/pull/358

It does appear that the 19-ea job has passed since then. That said, the
error was flaky, not a consistent failure, and thus the most recent
dozen-plus runs _before_ that PR had passed, too. Still, let's try this
and see how it goes.

This will hopefully wrap up our work on
https://github.com/jspecify/jspecify/issues/320.