docs: fail builds on invalid markdown links; update action
This commit updates our docusaurus config to throw an error on invalid
markdown links. I'm not certain that it will do much at all to help
catch issues like #346, but it will at least make our builds stricter in
general. We can disable this if it become to onerous.

This also refactors our Docs github workflow slightly to make it more
modular. We check builds on PRs to main and build and deploy only on a
push to main.