Stop running Error Prone.
Running it is good practice, but running it has also gotten more complex
now that Error Prone depends on JSpecify.

Luckily, there is only so much that we can mess up in the declarations
of the annotations (and only a little more than that once you include
the integration tests).

(Please don't think too hard about the fact that Error Prone is
currently producing legitimate InvalidBlockTag and
UnrecognisedJavadocTag warnings because we're using a JDK11 toolchain
for Javadoc at the moment....)

Fixes https://github.com/jspecify/jspecify-reference-checker/issues/202