Update rules for a change in `T*` substitution.
This PR implements the "`*/*/*/*`" (aka "`String*`") option from
https://github.com/jspecify/jspecify/issues/248.

As such, we should merge it only if we decide to go that way. If we
decide to go another way, then we probably need to [make a different
change to the the
spec](https://github.com/jspecify/jspecify/issues/623#issuecomment-2341517812):
We need to do _something_ for #248 as part of
https://github.com/jspecify/jspecify/issues/623.

Assuming that we do go that way:

I think this PR changes the rules such that we never say either "If `T`
has nullness operator `UNION_NULL`..." or "If `T` has nullness operator
`MINUS_NULL`...." Instead, checks for those nullness operators always
use the "worried" and "comfortable" rules. This is part of what I
conceptually like about both these new rules and the worried/comfortable
model, though it didn't fully sink in until I was writing this commit
message.

As part of simplifying the rules of substitution, this PR also...

Fixes https://github.com/jspecify/jspecify/issues/522

If we do not merge this PR, then we'll need to merge an alternative fix.