Treat Kotlin code as null-unmarked.
This rule is written such that a Kotlin class _would_ still be treated
as null-marked _if_ it were annotated with `@NullMarked`. That enables a
user to annotate it if desired, but more importantly, it allows a future
version of Kotlin to produce that annotation automatically.

Still, we are likely to want to revisit this rule, especially if Kotlin
begins generating `@NullMarked`. In particular, as [noted by
netdpb@](https://github.com/jspecify/jspecify/issues/611#issue-2489956831),
if a Kotlin class is nested inside a `@NullMarked` Kotlin class, then
the nested class should probably also be treated as null-marked. Under
this current rule, it probably isn't, since we would (IIUC) see a
`@kotlin.Metadata` annotation on the nested class before seeing the
`@NullMarked` annotation on the enclosing class.

We could try to do better, but this PR's approach has the virtue of
being very easy to implement.

(So this PR might not really "fix"
https://github.com/jspecify/jspecify/issues/611, but I propose that it's
good enough for spec-1.0.0 purposes. We could probably even release spec
1.0.0 without any PR like this if we really wanted.)