Deem "`equals(Object o)`" in a record class to always have unspecified nullness.
I chose to write a rule directly about the _type usage_ that is the
parameter, rather than to write a rule that makes the whole method a
null-unmarked scope as I'd been discussing in the issue thread. I did
this so that at least the _implementation_ can remain null-marked if the
class is.

One arguable theoretical downside to that choice is that even the
parameter's nullness-would be unspecified even in the case of:

```
@NullMarked record Foo() {
  @NullMarked public boolean equals(Object o) { ... }
}
```

But it would _also_ be weird if the method-level `@NullMarked` _had_ an
effect, since it's otherwise redundant with the class-level annotation.
So I'm not worried about that.

It would be nice if there were an obvious way to have our cake and eat
it, too, when analyzing the _source code_ for a null-marked record that
declares `equals` explicitly and omits `@Nullable` from its parameter:

- We'd like that to be an error, since the rules otherwise say that the
  type is non-nullable. (This exception for record `equals` methods
  exists only because we can't distinguish compiler-generated `equals`
  methods from handwritten ones when looking at _bytecode_.)
- But we don't want _callers_ to treat the `equals` method differently
  based on whether they happen to be part of the same javac invocation
  as the record class (especially in the presence of incremental
  recompilation and sharded compilation).

This PR's solution seems good enough for now, especially since tools can
(and [do](https://errorprone.info/bugpattern/EqualsMissingNullable))
have special cases for suggesting `@Nullable` on `equals` parameters
independent of whether the parameter is currently considered to be
non-null or to have unspecified nullness. We can hope that someday Java
will [provide a way to distinguish compiler-generated `equals` methods
from handwritten ones when looking at
bytecode](https://github.com/jspecify/jspecify/issues/301#issuecomment-2012625990).

This PR fixes the easily fixable part of
https://github.com/jspecify/jspecify/issues/301. I am going to see if I
can subsequently easily add at least a non-normative comment about the
other considerations that we're aware of related to bytecode analysis.