Improve explanation of why "the `ImmutableList.Builder` rule" produces `MINUS_NULL`.
I'm actually wondering if we should just cut this discussion entirely.
It's strictly less important than the _main_ explanation of the
`ImmutableList.Builder` rule, and we may remove _that_ explanation in
favor of a link to a doc.

PLus, I think it would be reasonable to argue that "obviously" you use
`MINUS_NULL` because you're substituting for a non-nullable type
parameter. Thus, the explanation I'm giving is a more complicated
explanation of something that users would naturally already expect.

(As for the actual content of the PR: I noticed that I _think_ the
`Sequence` class has to be null-marked (and not _merely_ to have
`extends @Nullable Object` in its type-parameter declaration) in order
for the example to hold up after our decision in
https://github.com/jspecify/jspecify/issues/248. If not for
`@NullMarked`, then the type (sans `IL.B` rule) would be `Optional<E*>`,
and dereferencing `E*` would not be an error to a lenient checker. We
need for the type to be `Optional<E>` (again, sans `IL.B` rule) so that
a dereference would be an error. That way, we demonstrate that the
`IL.B` rule saves us from that false-positive error.)