Touch on the syntax for annotating arrays.
This builds on the work of
https://github.com/jspecify/jspecify/pull/592 in sorta kinda partially
addressing https://github.com/jspecify/jspecify/issues/580, though that
request is more about the user guide.

We do want to save most of this discussion for the user guide and
perhaps [other docs](https://github.com/jspecify/jspecify/issues/550),
but we also want to make clear that this kind of annotating is
_supported_, especially since we already call out the similar case of
varargs.

(@sdeleuze)