Make dereference example less potentially misleading. (#687)
To check whether an expression should be dereferenced, a tool should
check whether it is null-exclusive under every parameterization. (The
spec says as much elsewhere.) But the text here could be read as
suggesting that the tool need only check the nullness operator.

Checking only the nullness operator is sufficient as long as the type
isn't a type variable. Thus, the simple fix here is to restrict the
example to a type (here, `String`) that isn't a type variable.