Rewrite the section more significantly:
- Use `Comparable` instead of `ImmutableList.Builder`.
- Present the rule as the natural behavior for a non-null type
  parameter, not a "special case."
- Stop overanalyzing why we pick `MINUS_NULL`: The type is non-null, so
  we want to substitute the way we would for any other non-null type.

(On that final point, note also that, thanks to the way the rules are
currently structured, we don't have an obvious way to produce
`NO_CHANGE` as the _final_ output because all we're doing in _this_ rule
is selecting which operator to _apply_. So, if we want to ensure that we
end up with a non-null result, then we need to apply `MINUS_NULL`. If we
wanted to end up with `NO_CHANGE` (which we don't!), then we'd
_additionally_ need to restructure these rules, perhaps to replace the
_type argument_ with a type argument whose nullness operator has been
replaced by `NO_CHANGE`. My point in all this is just that I am
_continuing_ to overthink why we use `MINUS_NULL`, which is simply the
natural choice.)